+++ app/controllers/orders_informations_controller.rb
@@ -1,0 +1,27 @@
+class OrdersInformationsController < ApplicationController
+  before_action :set_item, only: [:index, :create]
+
+  def index
+    @order_information = OrderInformation.new
+  end
+
+  def create
+    @order_information = OrderInformation.new(order_params)
+    if @order_information.valid?
+      @order_information.save
+      return redirect_to root_path
+    else
+      render 'index'
+    end
+  end
+
+  private
+
+  def order_params
+    params.require(:order_information).permit(:postal_code, :area_id, :city, :address, :building, :phone).merge(user_id: current_user.id, item_id: params[:item_id])
+  end
+
+  def set_item
+    @item = Item.find(params[:item_id])
+  end
+end
