--- app/helpers/orders_helper.rb
@@ -1,2 +1,0 @@
-module OrdersHelper
-end
