+++ app/helpers/orders_informations_helper.rb
@@ -1,0 +1,2 @@
+module OrdersInformationsHelper
+end
