+++ app/views/items/show.html.erb
@@ -30,7 +30,7 @@
     <%= link_to "削除", item_path(@item.id), method: :delete, class:"item-destroy" %>
    <% else %>
     <%# 商品が売れていない場合はこちらを表示しましょう %>
-    <%= link_to "購入画面に進む", "#" ,class:"item-red-btn"%>
+    <%= link_to "購入画面に進む",  item_orders_informations_path(@item.id), class:"item-red-btn"%>
     <%# //商品が売れていない場合はこちらを表示しましょう %>
     <% end %>
     <% end %>
