+++ config/routes.rb
@@ -1,5 +1,7 @@
 Rails.application.routes.draw do
   devise_for :users
   root to: "items#index"
-  resources :items
+  resources :items do
+  resources :orders_informations, only:[:index, :create]
+end
 end
