+++ db/migrate/20210612135727_create_information.rb
@@ -2,7 +2,7 @@ class CreateInformation < ActiveRecord::Migration[6.0]
   def change
     create_table :information do |t|
       t.string     :postal_code,   null: false
-      t.integer    :prefecture_id, null: false
+      t.integer    :area_id, null: false
       t.string     :city,          null: false
       t.string     :address,       null: false
       t.string     :building,      null: false
