+++ db/schema.rb
@@ -35,7 +35,7 @@ ActiveRecord::Schema.define(version: 2021_06_12_135727) do
 
   create_table "information", options: "ENGINE=InnoDB DEFAULT CHARSET=utf8", force: :cascade do |t|
     t.string "postal_code", null: false
-    t.integer "prefecture_id", null: false
+    t.integer "area_id", null: false
     t.string "city", null: false
     t.string "address", null: false
     t.string "building", null: false
