--- spec/factories/information.rb
@@ -1,5 +1,0 @@
-FactoryBot.define do
-  factory :information do
-    
-  end
-end
