+++ spec/factories/order_informations.rb
@@ -1,0 +1,19 @@
+# FactoryBot.define do
+#   factory :order_information do
+
+#     postal_code {123-4567}
+
+#     area_id     {1}
+
+#     city        {横浜市緑区}
+
+#     address     {青山1-1-1}
+
+#     building    {}
+
+#     phone       {09012345678}
+
+#     association :user
+#     association :item
+#   end
+#   end
