--- spec/factories/orders.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :order do
-    association :user
-    association :item
-  end
-end
