--- spec/models/order_spec.rb
@@ -1,7 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Order, type: :model do
-  before do
-    @order = FactoryBot.build(:order)
-  end
-end
