+++ spec/requests/orders_informations_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "OrdersInformations", type: :request do
+
+end
