--- spec/requests/orders_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Orders", type: :request do
-
-end
